Make a gem of the script and add a simple spec
